modified PID again

+++ cg/src/cg/pid/eval_forces.cpp
@@ -27,6 +27,11 @@ void eval_forces::iter(bundle_expr<E> const &bundle) const {
 
   auto type = bundle.type();
   auto i1p = prev[i1], i1n = next[i1], i2p = prev[i2], i2n = next[i2];
+  if (i1p < 0 || i1n < 0 || i2p < 0 || i2n < 0)
+    return;
+
+  lambda const *bb_lam_1, *bb_lam_2;
+  sink_lj const *bb_lj;
 
   {
     auto i1_ = i1, i2_ = i1n, i3_ = i1p, i4_ = i2;
@@ -35,11 +40,14 @@ void eval_forces::iter(bundle_expr<E> const &bundle) const {
     auto rm = cross(rij, rkj), rn = cross(rkj, rkl);
     auto rm_ninv = norm_inv(rm), rn_ninv = norm_inv(rn),
          rkj_ninv = norm_inv(rkj);
+    if (rm_ninv > (real)10.0 || rn_ninv > (real)10.0)
+      return;
     auto rkj_n = norm(rkj);
     auto fi = rm * rkj_n * rm_ninv * rm_ninv;
     auto fl = -rn * rkj_n * rn_ninv * rn_ninv;
     auto df = (fi * dot(rij, rkj) - fl * dot(rkl, rkj)) * rkj_ninv * rkj_ninv;
-    auto fj = -fi + df, fk = -fl - df;
+    auto fj = -fi + df;
+    auto fk = -fl - df;
 
     dpsi1_dr1 = fi;
     dpsi1_dr1n = fj;
@@ -50,6 +58,19 @@ void eval_forces::iter(bundle_expr<E> const &bundle) const {
     psi1 = acos(cos_psi1);
     if (dot(rij, rn) < (real)0.0)
       psi1 = -psi1;
+
+    int m = 1;
+    if (i2 - i1 != 3 || m != 2) {
+      bb_lam_1 = &bb_plus_lam;
+      bb_lj = &bb_plus_lj;
+    } else {
+      bb_lam_1 = &bb_minus_lam;
+      bb_lj = &bb_minus_lj;
+    }
+    if (!bb_lam_1->supp(psi1) && (i2 - i1 != 3 || m != 1)) {
+      bb_lam_1 = &bb_minus_lam;
+      bb_lj = &bb_minus_lj;
+    }
   }
 
   {
@@ -59,11 +80,14 @@ void eval_forces::iter(bundle_expr<E> const &bundle) const {
     auto rm = cross(rij, rkj), rn = cross(rkj, rkl);
     auto rm_ninv = norm_inv(rm), rn_ninv = norm_inv(rn),
          rkj_ninv = norm_inv(rkj);
+    if (rm_ninv > (real)10.0 || rn_ninv > (real)10.0)
+      return;
     auto rkj_n = norm(rkj);
     auto fi = rm * rkj_n * rm_ninv * rm_ninv;
     auto fl = -rn * rkj_n * rn_ninv * rn_ninv;
     auto df = (fi * dot(rij, rkj) - fl * dot(rkl, rkj)) * rkj_ninv * rkj_ninv;
-    auto fj = -fi + df, fk = -fl - df;
+    auto fj = -fi + df;
+    auto fk = -fl - df;
 
     dpsi2_dr2 = fi;
     dpsi2_dr2n = fj;
@@ -74,35 +98,50 @@ void eval_forces::iter(bundle_expr<E> const &bundle) const {
     psi2 = acos(cos_psi2);
     if (dot(rij, rn) < (real)0.0)
       psi2 = -psi2;
+
+    int m = 2;
+    if (i2 - i1 != 3 || m != 2) {
+      bb_lam_2 = &bb_plus_lam;
+      bb_lj = &bb_plus_lj;
+    } else {
+      bb_lam_2 = &bb_minus_lam;
+      bb_lj = &bb_minus_lj;
+    }
+    if (!bb_lam_2->supp(psi2) && (i2 - i1 != 3 || m != 1)) {
+      bb_lam_2 = &bb_minus_lam;
+      bb_lj = &bb_minus_lj;
+    }
   }
 
   real dV_dpsi1 = 0.0f, dV_dpsi2 = 0.0f, dV_dr = 0.0f, V_ = 0.0f;
 
-  auto &bb_lam_1 = bb_plus_lam.supp(psi1) ? bb_plus_lam : bb_minus_lam;
-  auto &bb_lam_2 = bb_plus_lam.supp(psi2) ? bb_plus_lam : bb_minus_lam;
-  auto &bb_lj = bb_plus_lam.supp(psi2) ? bb_plus_lj : bb_minus_lj;
+  if (bb_lam_1->supp(psi1) && bb_lam_2->supp(psi2)) {
+    auto [lam1, deriv1] = (*bb_lam_1)(psi1);
+    auto [lam2, deriv2] = (*bb_lam_2)(psi2);
 
-  if (bb_lam_1.supp(psi1) && bb_lam_2.supp(psi2)) {
-    auto [lam1, deriv1] = bb_lam_1(psi1);
-    auto [lam2, deriv2] = bb_lam_2(psi2);
-    auto [lj_V, lj_dV_dr] = bb_lj(r12_rn);
+    if (lam1 * lam2 > (real)5e-5) {
+      auto [lj_V, lj_dV_dr] = (*bb_lj)(r12_n, r12_rn);
 
-    V_ += lam1 * lam2 * lj_V;
-    dV_dpsi1 += deriv1 * lam2 * lj_V;
-    dV_dpsi2 += deriv2 * lam1 * lj_V;
-    dV_dr += lam1 * lam2 * lj_dV_dr;
+      V_ += lam1 * lam2 * lj_V;
+      dV_dpsi1 += deriv1 * lam2 * lj_V;
+      dV_dpsi2 += deriv2 * lam1 * lj_V;
+      dV_dr += lam1 * lam2 * lj_dV_dr;
+    }
   }
 
   sink_lj ss_sink_lj = ss_ljs[type];
   if (ss_lam.supp(psi1) && ss_lam.supp(psi2)) {
     auto [lam1, deriv1] = ss_lam(psi1);
     auto [lam2, deriv2] = ss_lam(psi2);
-    auto [lj_V, lj_dV_dr] = ss_sink_lj(r12_n, r12_rn);
 
-    V_ += lam1 * lam2 * lj_V;
-    dV_dpsi1 += deriv1 * lam2 * lj_V;
-    dV_dpsi2 += deriv2 * lam1 * lj_V;
-    dV_dr += lam1 * lam2 * lj_dV_dr;
+    if (lam1 * lam2 > (real)5e-5 && ss_sink_lj.depth() > (real)5e-5) {
+      auto [lj_V, lj_dV_dr] = ss_sink_lj(r12_n, r12_rn);
+
+      V_ += lam1 * lam2 * lj_V;
+      dV_dpsi1 += deriv1 * lam2 * lj_V;
+      dV_dpsi2 += deriv2 * lam1 * lj_V;
+      dV_dr += lam1 * lam2 * lj_dV_dr;
+    }
   }
 
   *V += V_;
